Bluespace Patch Patch
Fixes the ability to use the thread on an item that already been made roomy, and makes the spool way smaller so you can put it into a box.

Also, gives the spool more uses so that it lasts the shift, and makes it so towels have the "roomy" flag innately so you can always use them to cover your bits in a pinch.