Do Not Clone
Adds a new trait for people who don't want to be cloned. Gives the medical bay something to do at least.

Also, moves a define to it's appropriate folder to prevent any potential future problems because for some reason it does matter where those defines are for certain procs. The new trait is in negative.dm instead of the hyper traits file, because it populates the list in that order and not based on cost.