Uniform armor changes
SS13 in the code had several armor values on uniforms that had values beyond the percentage of 10, things like 60% to 80% fire resist on a uniform which can stack, although the fire can destroy the uniform for example, it won't harm the player for the duration it's still in good condition, leaving to some strange encounters in fights. Ontop of this some uniforms for jobs such as roboticist, atmos tech have been given values better suit all the other uniforms, while heads of departments are given the 10% armor protection much like security uniforms, makes sense as a head has to be protected, while the captain gets a 10% all around protection.