Length times Size
This is more so a change to examine text and flavor text revolving around our size code we have in game, ultimately if one were to be 50% sprite size (Which is roughly three feet tall as we're doing 6ft for basic height reference), and sporting a rod that's just the same size in comparison. It should instead be half the amount based off sprite size, ending up being a foot and six inches versus three feet.