X-01 MultiPhase Energy Gun Buff
Increased the charge capacity of the X-01's blaster by 60% along with it's charge capacity, allowing 16 lethal shots, 40 disabler, or 16 stun shots. Origionally the stun energy cost was at a whole 200 charge reduction , now being 100 to represent the power in comparison to a tactical energy gun that spent half the amount of energy in comparison.