Fix TRAIT_HEAT not being removed correctly
I believe proc REMOVE_TRAIT wanted you to specify what kind of trait it is that you wanted to remove. Haven't checked the proc behavior yet but when investigating its use through the code it does hint strongly that this is the case. 
Tested it on a private server and it appears to successfully remove TRAIT_HEAT when I do a var check on a character.