Prevent Portallights from being deepfried?
Tested and confirmed working. This edit will prevent portallights from being deepfried.

I will leave the decision to either merge or reject this PR to the admin looking into this. This literally took me like five minutes so don't feel bad if you need to kick this if its not aligning to the vision of the server.