Virgo's 'Busy Space' - - - Simply does as it implies on the tin. Ports the 'Busy Space' modules and reworks them slightly to function on TG.
It does not broadcast to Common as it does on Virgo/Polaris. Instead, it functions similar to announcements, near identically in fact, having its own little broadcast. I'll be posting examples alongside this.

If it's a necessity to be broadcasting over Common, I'd have to fight with it a bit, but can probably do so without issue.
- - -