Fix borg flavortext
Ensures silicon flavortext is properly populated whenever a borg mind initializes by creating a new signal for it in mind.dm. Brings savefile code for silicon flavortext in line with standards for OOC and IC flavortext.
Tested and works upon round start, when borgified, or when an AI deploys to a shell.