Teleporter Balance Adjustment
---Why---

My personal gripes is that this is a mechanics (among a few others) that will immutably change your char for the rest of the round. To reverse this process is either jarring from a pure IC view (if they have a clone backup you just kill them + clone and act as if it never happened) or immensely tedious, bothering the xenobiologist to breed a very specific slime and make extracts, and even that might not have guaranteed results.

This is a furry rp server with rounds that routinely last 4 hours and longer. I think it's also safe to assume that the majority of people who play here do so to play a character that they have specifically created. Having a meme that turns you into a flyperson doesn't feel very compatible with these traits. Sure, it's a punishment yes, one that can be avoided with a simple button press, but in the same light, isn't that a bit heavy handed?

---Balance---

I wrote the values as an initial guesstimate. I am not going to pretend I know how to balance an alternative to the fly meme, but I think what I did is a suitable punishment.
you will receive a lot of clone damage, some rads and for each limb it's a 50/50 on whether you loose it, including the head.