Damage Balancing, Polish
Re-enables and fixes applying animations upon falling into the chasm, adds wind-up to the teleportation. Balances the damages the Tether applies, including a small amount of time for the player to realize their mistake. Reworks the objects' coding for the center for some edge cases, or for if density is turned back on.