Painkillers
Adds a milder painkiller in the form of ibuprofen, made with 3 bicardine, 1 carbon, 1 hydrogen. It has a high overdose threshold of 60 and mild side effects when ODed. Chemfridges start with 5 pills of 15 units each, as do medical vendors. Borgs can use it by default, and borgs can now use morphine when upgraded for painkilling purposes.